+++ src/main/java/com/strangegrotto/wealthdraft/assets/definition/Asset.java
@@ -1,13 +1,22 @@
 package com.strangegrotto.wealthdraft.assets.definition;
 
+import com.fasterxml.jackson.annotation.JsonSubTypes;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.strangegrotto.wealthdraft.assetimpls.AssetTypeTagValue;
+import com.strangegrotto.wealthdraft.assetimpls.bankaccount.BankAccountAsset;
 import com.strangegrotto.wealthdraft.assets.temporal.AssetChange;
 import com.strangegrotto.wealthdraft.assets.temporal.AssetSnapshot;
 import com.strangegrotto.wealthdraft.errors.ValOrGerr;
 
 import java.util.Map;
 
-@JsonDeserialize(as = AbstractAsset.class)
+@JsonTypeInfo(
+        use = JsonTypeInfo.Id.NAME,
+        property = "type")
+@JsonSubTypes(
+        @JsonSubTypes.Type(value = BankAccountAsset.class, name = "BANK_ACCOUNT")
+)
 public interface Asset {
     String getName();
 
@@ -15,5 +24,9 @@ public interface Asset {
 
     Class<? extends AssetSnapshot> getSnapshotType();
 
-    ValOrGerr<Map<AssetTag, String>> getTags();
+    AssetTypeTagValue getAssetTypeTagValue();
+
+    Map<String, String> getAssetSpecificTags();
+
+    Map<String, String> getCustomTags();
 }