+++ src/main/java/com/strangegrotto/wealthdraft/assets/definition/AssetTag.java
@@ -1,5 +1,0 @@
-package com.strangegrotto.wealthdraft.assets.definition;
-
-public interface AssetTag {
-    String getName();
-}