+++ src/main/java/com/strangegrotto/wealthdraft/assets/definition/CustomTagDefinition.java
@@ -2,7 +2,6 @@ package com.strangegrotto.wealthdraft.assets.definition;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
-import com.strangegrotto.wealthdraft.assets.ImmCustomTagDefinition;
 import org.immutables.value.Value;
 
 import java.util.Set;
@@ -14,5 +13,6 @@ public interface CustomTagDefinition {
     // TODO Implement default values
 
     // An empty set means all values are allowed
+    @Value.Parameter
     Set<String> getAllowedValues();
 }