+++ src/main/java/com/strangegrotto/wealthdraft/assets/temporal/AssetParameterChangeDeserializer.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
-import com.strangegrotto.wealthdraft.assets.ImmAssetParameterChange;
 
 import java.io.IOException;
 import java.math.BigDecimal;