+++ src/main/java/com/strangegrotto/wealthdraft/networth/history/AssetsHistory.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.google.common.base.Preconditions;
 import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
 import com.strangegrotto.wealthdraft.assets.definition.Asset;
-import com.strangegrotto.wealthdraft.assets.definition.CustomTagDefinition;
+import com.strangegrotto.wealthdraft.assets.definition.AssetDefinitions;
 import com.strangegrotto.wealthdraft.assets.temporal.AssetSnapshot;
 import org.immutables.value.Value;
 
@@ -13,13 +13,10 @@ import java.util.Map;
 
 @WealthdraftImmutableStyle
 @Value.Immutable
-@JsonDeserialize(using = AssetsWithHistoryDeserializer.class)
-public interface AssetsWithHistory {
+public interface AssetsHistory {
+
     @Value.Parameter
     Map<String, Asset> getAssets();
-
-    // The user will need to declare the custom tags they want to use, along with default values
-    Map<String, CustomTagDefinition> getCustomTags();
 
     @Value.Parameter
     Map<String, Map<LocalDate, AssetSnapshot>> getHistory();
@@ -27,20 +24,24 @@ public interface AssetsWithHistory {
     @Value.Check
     default void check() {
         LocalDate today = LocalDate.now();
+        var assets = getAssets();
+        var history = getHistory();
 
-        // Verify that all assets in history are in the asset list
-        for (String assetId : getHistory().keySet()) {
+        // Verify that all assets in the history are in the asset list
+        for (String assetId : history.keySet()) {
             Preconditions.checkState(
-                    getAssets().containsKey(assetId),
-                    "Asset ID %s appears in the history but isn't defined in the assets section",
-                    assetId);
+                    assets.containsKey(assetId),
+                    "Asset '%s' appears in the history but isn't defined in the assets list",
+                    assetId
+            );
 
             for (LocalDate date : getHistory().get(assetId).keySet()) {
                 Preconditions.checkState(
                         !date.isAfter(today),
-                        "Asset %s has a historical record with date %s, which is in the future",
+                        "Asset '%s' has a historical record with date '%s', which is in the future",
                         assetId,
-                        date);
+                        date
+                );
             }
         }
     }